tesseract/requirements.txt
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.10.0
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
cssselect2==0.4.1
cycler==0.11.0
defusedxml==0.7.1
h11==0.12.0
idna==3.3
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.3
opencv-python==4.5.4.58
outcome==1.1.0
Pillow==8.4.0
pycparser==2.20
pyOpenSSL==21.0.0
pyparsing==3.0.4
pytesseract==0.3.8
python-dateutil==2.8.2
requests==2.26.0
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
table-ocr==0.2.5
tinycss2==1.1.0
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
webencodings==0.5.1
wsproto==1.0.0
